Simplified code and Refactor Business Registration System
