+++ frontend/src/pages/business-registration-page/TemplateFrame.jsx
@@ -11,7 +11,7 @@ import AppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import ArrowBackRoundedIcon from '@mui/icons-material/ArrowBackRounded';
 import ToggleColorMode from './ToggleColorMode';
-import getSignUpTheme from './theme/getSignUpTheme';
+import getSignUpTheme from './theme/getBusinessRegistrationTheme';
 
 const StyledAppBar = styled(AppBar)(({ theme }) => ({
   position: 'relative',
