+++ frontend/src/pages/business-registration-page/getBusinessRegistrationTheme.jsx
