+++ frontend/src/pages/business-registration-page/theme/getBusinessRegistrationTheme.jsx
